1. AddNewSamuraiViaDisconnectedBattleObject(): Adding Many-to-many Ends on the Fly. Add new Samurai to existing battle with many-to-many relationship.
2. GetSamuraiWithBattles():
If we want to retrive a Samurai and all of the battle Samurai has fought in,
we would query the Samurai, and then Include the SamuraiBattle attached to the Samurai
and then Include the Battle that's attached to each one of those SamuraiBattles.
Samurai <Include> SamuraiBattle <ThenInclude> Battle
Note: sb.Battle will not come with intellisence, VS 2017 issue.

3. RemoveBattleFromSamurai():
Remove join between Samurai and battle (by using Ids) in connected mode.

4. RemoveBattleFromSamuraiWhenDisconnected():
Remove join between Samurai and battle (by using Ids) in disconnected mode.